Rename dataservices/server to dataservices/service to prevent collision
