+++ dataservices/service/context/standard.go
@@ -16,14 +16,14 @@ import (
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/deduplicator"
 	"github.com/tidepool-org/platform/data/store"
-	"github.com/tidepool-org/platform/dataservices/server"
-	"github.com/tidepool-org/platform/service"
+	"github.com/tidepool-org/platform/dataservices/service"
+	commonService "github.com/tidepool-org/platform/service"
 	"github.com/tidepool-org/platform/service/context"
 	"github.com/tidepool-org/platform/userservices/client"
 )
 
 type Standard struct {
-	service.Context
+	commonService.Context
 	dataFactory             data.Factory
 	dataStoreSession        store.Session
 	dataDeduplicatorFactory deduplicator.Factory
@@ -31,7 +31,7 @@ type Standard struct {
 	requestUserID           string
 }
 
-func WithContext(dataFactory data.Factory, dataStore store.Store, dataDeduplicatorFactory deduplicator.Factory, userServicesClient client.Client, handler server.HandlerFunc) rest.HandlerFunc {
+func WithContext(dataFactory data.Factory, dataStore store.Store, dataDeduplicatorFactory deduplicator.Factory, userServicesClient client.Client, handler service.HandlerFunc) rest.HandlerFunc {
 	return func(response rest.ResponseWriter, request *rest.Request) {
 		context, err := context.NewStandard(response, request)
 		if err != nil {
